refactor : payment 관련 기능들을 패캐지화
