--- src/main/java/org/jungppo/bambooforest/entity/type/PaymentStatusType.java
@@ -1,6 +1,0 @@
-package org.jungppo.bambooforest.entity.type;
-
-public enum PaymentStatusType {
-	PENDING,
-	COMPLETED,
-}
