+++ src/main/java/org/jungppo/bambooforest/member/domain/RoleType.java
@@ -1,4 +1,4 @@
-package org.jungppo.bambooforest.entity.type;
+package org.jungppo.bambooforest.member.domain;
 
 public enum RoleType {
     ROLE_USER, ROLE_ADMIN
