+++ src/main/java/org/jungppo/bambooforest/member/dto/PaymentSetupResponse.java
@@ -1,8 +1,7 @@
-package org.jungppo.bambooforest.dto.payment;
+package org.jungppo.bambooforest.member.dto;
 
 import java.math.BigDecimal;
 import java.util.UUID;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PaymentSetupResponse {
-	private UUID paymentId;
-	private BigDecimal price;
+    private UUID paymentId;
+    private BigDecimal price;
 }
