+++ src/main/java/org/jungppo/bambooforest/member/presentation/MemberController.java
@@ -1,8 +1,9 @@
 package org.jungppo.bambooforest.member.presentation;
 
-import static org.springframework.http.HttpHeaders.*;
-import static org.springframework.http.HttpStatus.*;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.HttpStatus.CREATED;
 
+import lombok.RequiredArgsConstructor;
 import org.jungppo.bambooforest.member.dto.JwtDto;
 import org.jungppo.bambooforest.member.dto.MemberDto;
 import org.jungppo.bambooforest.member.service.MemberService;
@@ -17,36 +18,34 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import lombok.RequiredArgsConstructor;
-
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/members")
 public class MemberController {
 
-	private final MemberService memberService;
+    private final MemberService memberService;
 
-	@PostMapping("/logout")
-	public ResponseEntity<Void> logout(@AuthenticationPrincipal CustomOAuth2User customOAuth2User) {
-		memberService.logout(customOAuth2User);
-		return ResponseEntity.noContent().build();
-	}
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(@AuthenticationPrincipal CustomOAuth2User customOAuth2User) {
+        memberService.logout(customOAuth2User);
+        return ResponseEntity.noContent().build();
+    }
 
-	@GetMapping("/profile")
-	public ResponseEntity<MemberDto> getProfile(
-		@AuthenticationPrincipal CustomOAuth2User customOAuth2User) {
-		MemberDto memberDto = memberService.getProfile(customOAuth2User);
-		return ResponseEntity.ok().body(memberDto);
-	}
+    @GetMapping("/profile")
+    public ResponseEntity<MemberDto> getProfile(
+            @AuthenticationPrincipal CustomOAuth2User customOAuth2User) {
+        MemberDto memberDto = memberService.getProfile(customOAuth2User);
+        return ResponseEntity.ok().body(memberDto);
+    }
 
-	@PostMapping("/reissuance")
-	public ResponseEntity<JwtDto> reissuanceToken(
-		@RequestHeader(value = AUTHORIZATION) String refreshToken) {
-		try {
-			JwtDto jwtDto = memberService.reissuanceToken(refreshToken);
-			return ResponseEntity.status(CREATED).body(jwtDto);
-		} catch (InvalidRefreshTokenException e) {   // TODO. 함수형 인터페이스를 이용한 Refactoring
-			throw new RefreshTokenFailureException();
-		}
-	}
+    @PostMapping("/reissuance")
+    public ResponseEntity<JwtDto> reissuanceToken(
+            @RequestHeader(value = AUTHORIZATION) String refreshToken) {
+        try {
+            JwtDto jwtDto = memberService.reissuanceToken(refreshToken);
+            return ResponseEntity.status(CREATED).body(jwtDto);
+        } catch (InvalidRefreshTokenException e) {   // TODO. 함수형 인터페이스를 이용한 Refactoring
+            throw new RefreshTokenFailureException();
+        }
+    }
 }
