+++ src/main/java/org/jungppo/bambooforest/payment/domain/PaymentStatusType.java
@@ -1,0 +1,6 @@
+package org.jungppo.bambooforest.payment.domain;
+
+public enum PaymentStatusType {
+    PENDING,
+    COMPLETED,
+}
