+++ src/main/java/org/jungppo/bambooforest/repository/payment/PaymentRepository.java
@@ -1,8 +1,7 @@
 package org.jungppo.bambooforest.repository.payment;
 
 import java.util.UUID;
-
-import org.jungppo.bambooforest.entity.payment.PaymentEntity;
+import org.jungppo.bambooforest.payment.domain.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<PaymentEntity, UUID> {
