+++ src/main/java/org/jungppo/bambooforest/security/oauth2/response/OAuth2Response.java
@@ -1,10 +1,13 @@
 package org.jungppo.bambooforest.security.oauth2.response;
 
-import org.jungppo.bambooforest.entity.type.OAuth2Type;
+import org.jungppo.bambooforest.member.domain.OAuth2Type;
 
 public interface OAuth2Response {
     OAuth2Type getProvider();
+
     String getProviderId();
+
     String getName();
+
     String getProfileImage();
 }
